refactor: one declaration for the room / runtime / work-product wire contracts
#283 gave the chat surface a single declaration of its wire types after
finding five shapes spelled out ten times. The same duplication was still
in place for four other surfaces, even though the mechanism to fix it
already exists on both ends — `@beevibe/api/views/types` is a subpath
export web already re-exports fifteen DTOs through.

Fourteen shapes, twenty-eight declarations:

  the room message      MessageReply (api/routes/room.ts, not even
                        exported) + RoomMessage (web/lib/api/client.ts)
  the room row          core's Room, re-served with Dates; web declares
                        its own string-dated copy
  RoomMemberDetail,     inferred api-side from an object literal;
  the typing indicator   hand-written web-side
  RoomDetail            the envelope tying those together
  AddresseeReason       named in api; the union written out longhand at
                        web's sendMessage call site
  RuntimePanelEntry,    exported from api/routes/runtimes.ts AND declared
  DaemonPanelEntry,     again in web/lib/api/client.ts
  RuntimesListResponse
  WorkProductDetail     api/views/work-product.ts + web, with core's
                        WorkProductType union inlined on the web side

The runtime-panel copies had already drifted. The api sends
`cli_version: string | null` and `last_heartbeat: string | null`; web
declared both `?: string`, so `=== undefined` reads as the "never beat"
check while the server sends null. `capabilities` and `created_at` were
missing from web's copy entirely, and `device_name` was optional there
against a non-null field — the runtimes page still carries a
`?? "this daemon"` fallback for a value that cannot be absent. Pointing
web at the api's declaration surfaced the mismatch as two typecheck
errors, fixed in the two local RuntimeOption types.

WorkProductDetail's inlined `type` union is the RepoCardSource problem
from #283 verbatim: a tenth work-product type added to core would leave
the detail page's renderer silently unable to name it.

Promote all fourteen to views/types.ts and have both halves read them.
The api side now type-checks its own responses against the contract
rather than emitting whatever the object literal inferred: `toRoomSummary`
projects core's Date-bearing Room to the ISO-string wire shape (identical
JSON — Express was already stringifying these Dates), and the detail
handler builds a declared `RoomDetail`. `RoomDetail.typing` stays optional
because the web reads this shape out of a react-query cache that can hold
a response from a build predating the field.

Also folds daemon's `BatchedEvent` into `RuntimeEventInput` from
core/domain/daemon-protocol.ts — same four fields, and the protocol module
is where the rest of the /runtime/* contract already lives. The batcher's
POST body is now the declared `RuntimeEventsRequest`.

typecheck, lint and build clean across all six packages; web builds via
`next build`. api 820 tests, core 609, web 203, daemon 156, sandbox 109
and scheduler 10 pass — remaining failures are the documented
no-Postgres / no-provider-key baseline, unchanged from main.
