+++ packages/web/lib/api/types.ts
@@ -21,4 +21,15 @@ export type {
   CoreSnapshotRow,
   AgentRatioRow,
   BeforeAfterData,
+  AddresseeReason,
+  DaemonPanelEntry,
+  RoomDetail,
+  RoomMemberDetail,
+  RoomMessageDetail,
+  RoomSummary,
+  RoomTypingIndicator,
+  RoomTypingStep,
+  RuntimePanelEntry,
+  RuntimesListResponse,
+  WorkProductDetail,
 } from "@beevibe/api/views/types";
